🧹 organized auth files
placed auth hooks and context in their corresponding files and folders.
