+++ front/src/App.tsx
@@ -3,7 +3,7 @@ import { Outlet } from "react-router-dom";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
 import { createContext } from "react";
-import { AuthStatus } from ".";
+import { AuthStatus } from "./routes/Auth/AuthStatus";
 
 let LoginStatus = {
   islogged: false,
