+++ front/src/globals/contexts.tsx
@@ -1,0 +1,8 @@
+import { createContext, useContext } from "react";
+import { AuthContextType } from "./Interfaces";
+
+export let AuthContext = createContext<AuthContextType>(null!);
+
+export function useAuth() {
+  return useContext(AuthContext);
+}
