+++ front/src/routes/Chat.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import "./Chat.css";
 import { socket } from "../App";
-import { useAuth } from "..";
+import { useAuth } from "../globals/contexts";
 
 export default function Chat() {
   const id = useRef(0);
